requirements.txt
fastapi
uvicorn
pydantic
streamlit
requests
streamlit_autorefresh
openai
pandas
python-multipart
python-dotenv
sqlalchemy
asyncpg
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]


